Fix DOM objects to appear on click
