+++ src/index.js
@@ -2,7 +2,7 @@
 import 'bootstrap';
 import './css/main.scss';
 import todoItems from './js/todoItem';
-import domItems from './js/domObjects';
+import domItems from './js/listItems';
 
 
 let projectList;
@@ -13,7 +13,7 @@ if (localStorage.getItem('projectlist')) {
   projectList = todoItems.List;
   projectList.name = 'Projects';
 }
-// localStorage.setItem('projectlist', JSON.stringify(projectList));
+localStorage.setItem('projectlist', JSON.stringify(projectList));
 
 const sideBar = document.getElementById('left-sidebar');
 sideBar.appendChild(domItems.ProjectList(projectList));
